Added support for Canadian locales (ENCAN, FRCAN)
Support for two more locales added.  The 'English - Canada' (ENCAN) locale, and the "French - Canada' (FRCAN) locale.  Works exactly the same as before, with just two new options in the locale pulldown, and match code definitions for ENCAN and FRCAN, as well as the definitions for the original five locales are correctly synced in the Definition pulldown.  So no new or changed documentation is necessary.